requirements.txt
pip~=23.1.2
emoji~=2.2.0
attrs~=23.1.0
build~=0.10.0
cryptography~=41.0.1
Jinja2~=3.1.2
rsa~=4.9
pyasn1~=0.5.0
MarkupSafe~=2.1.3
Werkzeug~=2.3.6
click~=8.1.3
blinker~=1.6.2
itsdangerous~=2.1.2
six~=1.16.0
cachetools~=5.3.0
certifi~=2023.5.7
urllib3~=2.0.3
requests~=2.31.0
aiohttp~=3.8.4
cffi~=1.15.1
google-api-python-client~=2.88.0
pyparsing~=3.0.9
httplib2~=0.22.0
oauthlib~=3.2.2
idna~=3.4
setuptools~=65.5.0
packaging~=23.1
uritemplate~=4.1.1
legacy~=0.1.6
Flask~=2.2.5
gunicorn